move ics26_routing dispatch tests
